🎻 Adds tuned yaw controller, PID class

+++ lib/common/utils.cpp
@@ -3,7 +3,7 @@
 #include "utils.h"
 
 
-float mapfloat(long x, long in_min, long in_max, long out_min, long out_max) {
+float mapfloat(float x, float in_min, float in_max, float out_min, float out_max) {
   return (float)(x - in_min) * (out_max - out_min) / (float)(in_max - in_min) + out_min;
 }
