+++ lib/common/utils.h
@@ -6,7 +6,7 @@
 String getResetReasons();
 String timeAgo(int seconds);
 
-float mapfloat(long x, long in_min, long in_max, long out_min, long out_max);
+float mapfloat(float x, float in_min, float in_max, float out_min, float out_max);
 
 extern const char* adafruitRootCert;
